refactor: extract job5156 detail functions into lib module (Phase 2b)
Move 13 job5156 detail page extraction functions from content.ts into
lib/job5156-extractor.ts using the same createJob5156Extractor(deps)
DI factory pattern used by seek-extractor. Reduces content.ts by 845
lines (44 insertions, 845 deletions).

Includes two bug fixes found during per-function verification:
- buildJob5156DetailWorkHistoryItemFromApi: restored inline
  descriptionLines.join("\n") behavior (extractor had description
  which omits companyMeta/department/reasonText)
- buildJob5156EducationItemFromApi: restored inline implementation
  with degree/speciality/description fields (extractor was a
  simplified version missing these fields)
